aflow/pyproject.toml
[project]
name = "aflow"
version = "0.1.0"
description = ""
authors = [
    {name = "sungminna",email = "sungmin.na330@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pytest-playwright (>=0.7.0,<0.8.0)",
    "langfuse (>=3.0.7,<4.0.0)",
    "deepeval (>=3.2.1,<4.0.0)",
    "openai (>=1.93.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
